Handle process exits of workers
It is possible that a worker will exit with EXIT signal if it spawns a
linked process, which dies. In order to handle this the parent of the
Executor should trap exits and handle EXIT messages.

Unfortunately there is no easy way to render stack traces in this
situation.

In the latest faktory spec the number of retries is specified with a
"retry" key, not "retries" as it used to be. Updated the code and the
documentation.